requirements.txt
assemblyai>=0.44.3
crewai[google-genai]>=0.193.2
crewai-tools>=0.73.1
fastembed>=0.7.3
fastapi>=0.109.0
firecrawl-py>=4.3.6
google-generativeai>=0.8.5
kokoro>=0.9.4
pymilvus[milvus-lite]>=2.6.2
pymupdf>=1.26.4
python-dotenv>=1.1.1
python-multipart>=0.0.6
soundfile>=0.12.1
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
uvicorn[standard]>=0.27.0
yt-dlp>=2024.12.13
zep-cloud>=3.4.3
zep-crewai>=1.1.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
